requirements.txt
absl-py
numpy
torch
transformers
datasets
modelscope
adaseq
